update:重写用户成长路线

+++ service-all/navigation-backend-service/src/main/java/com/example/navigation/entity/User.java
@@ -11,7 +11,7 @@ public class User {
     @Id
     @Column(name = "user_id", length = 50)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long userId;                                 // 用户ID，自增主键
+    private Integer userId;                                 // 用户ID，自增主键
 
     @Column(unique = true, nullable = false, length = 50)
     private String username;                         // 用户名，唯一，不能为空
